SpawnPoint in a separate header
